Add support for pager and editor.
Use the pager for both the SHOW and LIST command.
The pager can be set for all commands via the core.pager configuration
option. It can also be set per-command via the pager.<command> options.
Setting pager.<command> to the empty value will disable the pager for
that command. Setting core.pager to the empty value will disable the
pager for every commands.